:recycle: store 별도 분리
